key editing feature and reworked hex view
